[NU-2414] Update Flink to 2.2, switch to our Chill fork with Kryo 5.6.2
