+++ version.sbt
@@ -1,1 +1,1 @@
-ThisBuild / version := "1.2.2-SNAPSHOT"
+ThisBuild / version := "2.0.0.3-SNAPSHOT"
